requirements.txt
google-generativeai>=0.3.0
python-dotenv
watchdog
mcp
